Extended CLI driver to accept operator declarations before starting to parse.
This is far better than exposing a set of hard-coded operator declarations.

Currently all operator tables start empty and all operator declarations need to
precede any input lines fed to the parser.

Rudimentary input scrubbing was implemented, but it doesn't cover all cases that
are not supported by the parser.

Moreover in case of a syntax error in the input, "Enforcement failed" and the
parser source line were the error was detected are shown instead of a
descriptive message.

I think all currently implemented parser functions are exposed by the driver.

--HG--
branch : feature