Added version tags starting with v0.0.0 and increasing for features and bugfixes
The idea is to comply with [SemVer 2.0.0](https://semver.org/), but I have no
stable public API, so major version gets 0 and the standard no longer
prescribes a specific way to increase minor and patch versions. So I increase
minor version when significant new functionality is added and patch version
when bugs are fixed.

--HG--
branch : administrative