Documented the minimum compiler version requirements to build the project.
Use of `multiwayUnion` in `dopParser.d` required a somewhat more recent version
of the standard library than the rest of the code as it was introduced around 3
years ago. I determined the minimum release version that supports it for `dmd`,
`ldc` and `gdc` and documented these in `dub.sdl`, while adding a note in
`readme.md` to refer to the former.

Documenting the version requirements in the package description has the added
benefit that recent versions of `dub` will fail early with a descriptive message
if they detect an attempt to build using an unsuitable toolchain. (e.g. `dub`
v1.21 does detect that, but v1.5 doesn't)

Determining the minimum required version was done using a binary search over the
version space of each compiler and attempting to build, as I was unable to find
any reference to the function or it's module in the release notes.
I used [install.sh](https://dlang.org/install.html) to install multiple compiler
versions side by side.

For `gdc`, the GitHub repository didn't contain any recent releases and
`install.sh` couldn't install any either, so I had to resort to a `docker`
container and a more recent version of `Ubuntu`. At the end I was able to
determine that `gdc` v5.4.0 is missing the required standard library function
but v8.4.0 doesn't. It's possible that some version between those can be used to
build the project as well, but given the difficulty of getting such a version,
requiring v8.4.0 seems reasonable at this time.

--HG--
branch : documentation