Renamed '++' operator to '%' in negative unit tests so that lines align better.
All other tokens used in these unit tests are one character long and the
purpose of these tests is to test the parser, not the scanner.

Covering all the combinations in source code requires quite a bit of copy-paste-
modify and modification is a lot easier when you can use the column selection
mode of the editor.

Only negative unit tests were changed. Positive ones were left as they were.

--HG--
branch : refactoring