Added negative unit tests involving juxtaposition for the dop-parser.
All operator definitions used in the sans-juxtaposition test cases were left
intact and the juxtaposition operator was added to them. In addition all the
test inputs remained the same and just the outputs were updated. In the majority
of cases even the outputs could be kept the same and ~1373 of them were moved to
a module-level `common_test_cases` array that was used in both `unittest`
blocks.

Approximately 4671 new test cases were added in total.

Remaining work includes:

- Fill in the correct output in tests where only success of the parsing is
  tested.
- Move test cases to separate files and further organize them.
- Prove certain uncovered lines are unreachable and remove them.
- Generalize support for confix and distfix operators and include negative tests
  for them as well.

--HG--
branch : testing