Added the dockerfile used to test toolchain requirements to version control.
A few weeks ago I created a dockerfile to build an Ubuntu-based image to more
easily test building with multiple different types and versions of D compilers.
It was used to test different environments and configurations before attempting
to use them with TravisCI.

Although useful, this version of the dockerfile should be considered a draft and
images built from it aren't intended to be pushed to DockerHub. More work is
needed to refine it and ideally it should setup only the build environment
leaving the actual build to be initiated by external build automation scripts.

Original Timestamp:

2020-05-23 02:09:23 +0300 dockerfile

--HG--
branch : build