Addressed an issue where the expected error message wouldn't be produced.
Before the fix, invalid expressions where a separator would appear just before
a terminator or end-of-input produced either an assertion error (if enabled) or
empty output. Instead a descriptive error message is now communicated to the
caller as an exception.

e.g. for an expression like "( ) ," you'll now get the message "A separator must
be followed by an operable expression!"

This was discovered while adding unit tests.

Also created a number of additional constants for error messages that
`parseInfixExpression` could produce and used them to replace the corresponding
string literals in the latter.

--HG--
branch : bugfix