Refactored logic in processToken to better match the algorithm in readme.md.
When comparing one non-operator token with the next, 4 possibilities are laid
out in `readme.md`, but only three where present in `processToken` while the 4th
was left to be checked later in `dispatchToken`. Now all conditions are together
and it's easier to see that all combinations are covered.

Also:

- Moved a few enforcements to `processToken` and left only assertions in their
  place in `dispatchToken`. Ideally, and eventually, all input scrubbing should
  be done in one place.
- Generalized the error message about two consecutive operands and
  juxtaposition operator being undefined and removed specific error when 2nd
  operable expression starts with an initiator. I can increase error granularity
  again later when code won't be based on conditionals any more and
  cyclomatic complexity won't be as high.

--HG--
branch : refactoring