Converted dop_parser.d into a package moving unit tests into separate files.
`dop_parser.d` was nearing 2k lines of code and it was high time it was split
into multiple files. All production code moved to one source file for now, but
unit tests were split into broad categories and separated into multiple files.

Also:

- Removed the `sourcePaths` and `importPaths` settings again and moved
  `test_utilities.d` back into the `source` folder.
- Updated column widths in `test_utilities.d` to account for the longer file
  names and paths.
- Gave a number of constants `package` access to make them available to tests.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

It appears that both D and DUB are designed under the assumption that tests for
the contents of a module will physically live inside the same file as the code
they are testing:

- It's really easy to use private constants from unit tests that reside into the
  same file, but there is no way to give access to those symbols to unit tests
  residing in other files.
- There is no separate build step and/or configuration that would allow unit
  tests access to source files and libraries not available to production code.
- There is no way to specify certain paths (e.g. for unit tests) that should be
  excluded from code coverage calculations. Since in D there is a lexical
  distinction between test and production code, this shouldn't be necessary but
  the current state of affairs is that there is no way to exclude test code
  from code coverage calculations.

The interplay between files and directories and modules and packages, also poses
a number of limitations in organizing test and production code effectively.
Finding a satisfactory option requires some thought and deeper knowledge of the
language. I had to revise the "Scaling Up" chapter from the "The D Programming
Language" book, although the newer "Package Module" feature from
[online reference](https://dlang.org/spec/module.html#package-module) ended up
giving the base for this solution.

Ideally I'd want access to at least one of two equally preferred organizational
schemes:

- For each production file `source/path/to/foo.d` to have multiple test
  files `test_group_1.d`, `test_group_2.d`, etc. under `test/path/to/foo`
  directory. Test directory foo can be replaced by a test file foo.d if there
  aren't enough tests to justify a whole directory.
- For each file `source/path/to/foo.d` that may initially contain both
  production and test code, once it grows too big, I should be able to replace
  it with a directory that contains a file for production code and one or more
  files for test code. The clients of the initial file shouldn't have to change.

In both cases, test files should be able to import the production file and other
test files to get access to common symbols if needed.

Since there is no way to exploit the separation of production and test code into
two parallel hierarchies as is done in Java with Maven, I decided to let test
code live under the same root as production code for now.

For the time being, the process will be as follows. Initially unit tests will be
written inside the same file as the production code they are testing. Once the
test suite grows enough, the file will be moved under a directory with the same
name that will also contain the test files. The structure will look like this:

    source
    ├── path
    │   └── to
    │       └── production
    │           ├── package.d
    │           ├── production.d
    │           └── tests
    │               ├── test_1.d
    │               ├── test_2.d
    │               └── test_3.d

`package.d` will be publicly importing `production.d` but will otherwise be
empty. It's needed so I don't end up with all heavily tested files being named
`package.d`.

--HG--
branch : refactoring