Modified copyright notice to make it compatible with DDoc.
[D's Documentation Generator](https://dlang.org/spec/ddoc.html) defines a
special `Copyright:` section for documenting `module` declarations. Adopting the
specified syntax will enable tools to extract the copyright notice and include
it in generated documentation for the files.

There are a number of different options for generating documentation from D
source files that aim to be compatible with the above specification and all have
slight differences in how they handle certain constructs:

- DDoc built into dmd.
  Used with `dub build --build=docs`.
  Wouldn't pick up the copyright section by default.
- [DDOX](https://github.com/rejectedsoftware/ddox).
  Used with `dub build --build=ddox`.
  Reflows text and removes links in "<...>".
- [harbored-mod](https://github.com/dlang-community/harbored-mod).
  Installed with `dub fetch harbored-mod; dub build harbored-mod`.
  Used with `path/to/hmod .`.

The `copyright` section in D documentation comments is only picked up when there
is a `module` declaration, so I added one to all remaining files.

Also replaced some tabs inside the copyright notices with single spaces.

--HG--
branch : documentation