First take on using D's safety attributes.
Annotated all modules with the `@safe` attribute although I had to use the
`@trusted` and `@system` attributes on some functions for the program to
compile.

Intuitively all code written so far in this project should be safe as I haven't
used pointer arithmetic, type punning or similar constructs, but I'm getting
some unexpected and not very helpful errors from some code unless I use lower
safety levels for it. e.g. the compiler complains about standard library
functions like `stdin.byLineCopy` and `object.keys`.

In addition to the above, catching `Throwable` isn't allowed in `@safe` code and
since I need it to report errors in tests, the `catch` in question had to be
moved out of the `runUnitTests` function and put into a separate function that
was declared `@trusted`. That should minimize the `@trusted` code.

More work is needed to understand how to make all code `@safe`.

--HG--
branch : enhancement