Updated dockerfile to account for bugs in `dmd` and `dub`.
I've been spending most of my time writing and running unit tests lately and I
had not run the application for a while. When I finally tried, I realised I
couldn't build it any more! The reason it couldn't be built was apparently a
compiler/linker bug that I reported [1], while the reason the problem went
unnoticed for so long was a package/build manager bug [2] that I also reported.

- Followed each `dub test` line with a `dub build` one as apparently `dub test`
  excludes `main.d` from the build and errors easily slip by [1].
- Started building with latest `dmd` version as well so problems with recent
  versions [2] won't go unnoticed in the future.

[1]: https://issues.dlang.org/show_bug.cgi?id=21071
[2]: https://github.com/dlang/dub/issues/1981

--HG--
branch : maintenance